Crate paralegal_spdg

This crate defines the program dependence graph (PDG) generated by Paralegal.

The top-level type is ProgramDescription. This type references multiple types defined within the Rust compiler such as MIR locations. To avoid requiring a rustc_private dependency on paralegal_spdg clients, we provide proxies in the rustc_proxies module for all Rustc types within the PDG.

Re-exports

pub use node_cluster::NodeCluster;

Modules

dot

Display SPDGs as dot graphs

node_cluster

Collections of nodes in a single controller

rustc 🔒

rustc_portable

Exports either rustc identifiers or their proxies depending on whether the rustc feature is enabled.

rustc_proxies

Proxies for Rustc types used within the PDG.

ser

Canonical serialziation to use. This is so that paralegal-flow and paralegal-policy agree on the format to use.

ser_defid_map 🔒

ser_defid_seq 🔒

ser_localdefid_map 🔒

tiny_bitset 🔒

tiny_bitset_pretty

Serialization that is readable. Serializes the set as a list of integers (that are set to one).

traverse

Utilities for traversing an SPDG

utils

Utility functions and structs

Structs

CallString

A location within the global call-graph.

DefInfo

Exported information from rustc about what sort of object a DefId points to.

DisplayNode

A structure with a Display implementation that shows information about a node index in a given graph.

DisplayPath

Provides a way to format rust paths

EdgeInfo

Metadata for an edge in the SPDGImpl

FunctionCallInfo

Metadata on a function call.

GlobalEdge

The global version of an edge that is tied to some specific entrypoint

GlobalLocation

A RichLocation within a specific point in a codebase.

GlobalNode

A globally identified node in an SPDG

GlobalNodeIter

Return type of IntoIterGlobalNodes::iter_global_nodes.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

Identifier

An identifier for any kind of object (functions, markers, etc.).

InstructionInfo

Information about an instruction represented in the PDG

MarkerAnnotation

A marker annotation and its refinements.

NodeInfo

Node metadata in the SPDGImpl

ProgramDescription

The annotated program dependence graph.

SPDG

A semantic PDG, e.g. a graph plus marker annotations

SPDGStats

Statistics about the code that produced an SPDG

ShortHash

Create a hash for this object that is no longer than six hex digits

SourceFile

An interned SourceFileInfo

SourceFileInfo

Information about a source file

Span

Encodes a source code location

SpanCoord

A “point” within a source file. Used to compose and compare spans.

TinyBitSet

A bit-set implemented with a u16, supporting domains up to 16 elements.

TypeDescription

Metadata about a type

Types

Holds TypeIds that were assigned to a node.

Enums

DefKind

Similar to DefKind in rustc but not the same!

EdgeKind

The type of an edge

InstructionKind

The type of instructions we may encounter

RichLocation

Extends a MIR body’s Location with Start (before the first instruction) and End (after all returns).

SourceUse

Additional information about the source of data.

TargetUse

Additional information about this mutation.

Constants

FLOW_GRAPH_OUT_NAME

Name of the file used for emitting the serialized ProgramDescription.

Traits

IntoIterGlobalNodes

This lets us be agnostic whether a primitive (such as flows_to) is called with a GlobalNode or &NodeCluster.

Functions

hash_pls

Calculate a hash for this object

Type Aliases

ControllerMap

Endpoints with their SPDGs

Endpoint

The types of identifiers that identify an entrypoint

Function

Identifiers for functions

Node

Local nodes in an SPDGImpl

SPDGImpl

The graph portion of an SPDG

TypeId

Identifiers for types

TypeInfoMap

information about each encountered type.