paralegal_spdg
Function allocative_visit_map_coerce_key

pub fn allocative_visit_map_coerce_key<'a, K, V>(
    map: &'a HashMap<K, V>,
    visitor: &mut Visitor<'a>,
)
where
    V: Allocative,

A function that is fit to be handed to #[allocative(visit = "...")] for a map where the key does not implement Allocative, but also has no attached heap data.

Uses SimpleSizedAllocativeWrapper under the hood, see its documentation for more information.