Struct paralegal_spdg::rustc::mir::traversal::Preorder

pub struct Preorder<'a, 'tcx> {
    body: &'a Body<'tcx>,
    visited: BitSet<BasicBlock>,
    worklist: Vec<BasicBlock, Global>,
    root_is_start_block: bool,
}

Preorder traversal of a graph.

Preorder traversal is when each node is visited after at least one of its predecessors. If you are familiar with some basic graph theory, then this performs a depth first search and returns nodes in order of discovery time.

        A
       / \
      /   \
     B     C
      \   /
       \ /
        D

A preorder traversal of this graph is either A B D C or A C D B

Fields
body: &'a Body<'tcx>visited: BitSet<BasicBlock>worklist: Vec<BasicBlock, Global>root_is_start_block: bool