paralegal_spdg
Struct GlobalEdge

pub struct GlobalEdge {
    pub(crate) index: EdgeIndex,
    pub(crate) controller_id: Endpoint,
}

The global version of an edge that is tied to some specific entrypoint

Fields
index: EdgeIndexcontroller_id: Endpoint
Implementations

impl GlobalEdge

pub fn controller_id(self) -> Endpoint

The id of the controller that this edge is located in