paralegal_spdg::tiny_bitset
Struct TinyBitSet

pub struct TinyBitSet(u16);

A bit-set implemented with a u16, supporting domains up to 16 elements.

Tuple Fields
0: u16
Implementations

impl TinyBitSet

pub fn new_empty() -> Self

Creates a new, empty bitset.

pub fn set(&mut self, index: u32)

Sets the indexth bit.

pub fn clear(&mut self, index: u32)

Unsets the indexth bit.

pub fn set_range(&mut self, range: Range<u32>)

Sets the ith to jth bits.

pub fn is_empty(self) -> bool

Is the set empty?

pub fn within_domain(self, index: u32) -> bool

Returns the domain size of the bitset.

pub fn count(self) -> u32

How many bits are set to true?

pub fn contains(self, index: u32) -> Option<bool>

Returns if the indexth bit is set.

pub fn is_set(self, index: u32) -> bool

Similar to Self::contains but in the case of an index outside the domain it returns false instead of None.

pub fn into_iter_set_in_domain(self) -> impl Iterator<Item = u32> + Clone

Iterate over all set indices that are within the domain (less than 16).

pub fn display_pretty(self) -> DisplayTinyBitSet

Create a struct with a pretty Display implementation for this tiny bit set.

pub fn intersection(self, other: Self) -> Self

A new bit set that contains elements that are both in self and other. Same as using &