paralegal_spdg
Module traverse

Utilities for traversing an SPDG

Enums

EdgeSelection

Which type of edges should be considered for a given traversal

Functions

bfs_iter 🔒

The current policy for this iterator is that it does not return the start nodes uness there is a cycle and a node is reachable that way.

generic_flows_to

A primitive that queries whether we can reach from one set of nodes to another

generic_influencees

Base function for implementing influencees

generic_influencers

Base function for implementing influencers