paralegal_spdg::traverse
Enum EdgeSelection

pub enum EdgeSelection {
    Data,
    Control,
    Both,
}

Which type of edges should be considered for a given traversal

Variants

Data

Consider only edges with crate::EdgeKind::Data

Control

Consider only edges with crate::EdgeKind::Control

Both

Consider both data and control flow edges in any combination

Implementations

impl EdgeSelection

pub const fn is_data(&self) -> bool

Returns true if the enum is EdgeSelection::Data otherwise false

pub const fn is_control(&self) -> bool

Returns true if the enum is EdgeSelection::Control otherwise false

pub const fn is_both(&self) -> bool

Returns true if the enum is EdgeSelection::Both otherwise false

impl EdgeSelection

pub fn use_control(self) -> bool

Does this selection admit edges of type crate::EdgeKind::Control

pub fn use_data(self) -> bool

Does this selection admit edges of type crate::EdgeKind::Data

pub fn conforms(self, kind: EdgeKind) -> bool

Is this edge kind admissible?

pub fn filter_graph<G: IntoEdgeReferences + Data<EdgeWeight = EdgeInfo>>( self, g: G, ) -> EdgeFiltered<G, fn(_: G::EdgeRef) -> bool>

Create a graph adaptor that implements this edge selection.