Module paralegal_spdg::utils

Utility functions and structs

Modules

serde_map_via_vec

Serialize a HashMap by converting it to a Vec, lifting restrictions on the types of permissible keys.

Structs

DisplayList

Has a Display implementation if the elements of the iterator inside have one. This will render them surrounded by [ brackets and separated by , comma and space

TruncatedHumanTime

A struct with a Display implementation taht renders a std::time::Duration in human readable form, similar to the humantime crate, but instead of rendering with arbitrary precision it only renders two “significant sections”, e.g. “2h 5min” or “2d 20h”. The sections are days (d), hours (h), minutes (min), seconds (s), miliseconds (ms), microseconds (μs) and nanoseconds (ns).

Functions

display_list

Display this iterator as a list

write_sep

Write all elements from it into the formatter fmt using f, separating them with sep