paralegal_spdg::node_cluster
Struct NodeCluster

pub struct NodeCluster {
    controller_id: Endpoint,
    nodes: Box<[Node]>,
}

A globally identified set of nodes that are all located in the same controller.

Sometimes it is more convenient to think about such a group instead of individual GlobalNodes

Fields
controller_id: Endpointnodes: Box<[Node]>
Implementations

impl NodeCluster

pub fn new( controller_id: Endpoint, nodes: impl IntoIterator<Item = Node>, ) -> Self

Create a new cluster. This for internal use.

pub fn iter(&self) -> Iter<'_>

Iterate nodes borrowing self

pub fn controller_id(&self) -> Endpoint

Controller that these nodes belong to

pub fn nodes(&self) -> &[Node]

Nodes in this cluster

pub fn try_from_iter(iter: impl IntoIterator<Item = GlobalNode>) -> Option<Self>

Attempt to collect an iterator of nodes into a cluster

Returns None if the iterator was empty or if two nodes did not have identical controller idโs