paralegal_spdg
Struct ProgramDescription

pub struct ProgramDescription {
    pub controllers: ControllerMap,
    pub type_info: TypeInfoMap,
    pub instruction_info: HashMap<GlobalLocation, InstructionInfo>,
    pub def_info: HashMap<DefId, DefInfo>,
    /* private fields */
}

The annotated program dependence graph.

Fields
controllers: ControllerMap

Entry points we analyzed and their PDGs

type_info: TypeInfoMap

Metadata about types

instruction_info: HashMap<GlobalLocation, InstructionInfo>

Metadata about the instructions that are executed at all program locations we know about.

def_info: HashMap<DefId, DefInfo>

Metadata about the DefIds

Implementations

impl ProgramDescription

pub fn canonical_write(&self, path: impl AsRef<Path>) -> Result<()>

Write self using the configured serialization format

pub fn canonical_read(path: impl AsRef<Path>) -> Result<Self>

Read self using the configured serialization format

impl ProgramDescription

pub fn all_nodes(&self) -> HashSet<GlobalNode>

Gather all data sources that are mentioned in this program description.

Essentially just self.controllers.flat_map(|c| c.keys())

pub fn all_call_sites(&self) -> HashSet<CallString>

Gather all CallStrings that are mentioned in this program description.