paralegal_spdg
Struct SPDGStats

pub struct SPDGStats {
    pub unique_locs: u32,
    pub unique_functions: u32,
    pub analyzed_locs: u32,
    pub analyzed_functions: u32,
    pub inlinings_performed: u32,
    pub construction_time: Duration,
    pub conversion_time: Duration,
}

Statistics about the code that produced an SPDG

Fields
unique_locs: u32

The number of unique lines of code we generated a PDG for. This means MIR bodies without considering monomorphization

unique_functions: u32

The number of unique functions that became part of the PDG. Corresponds to Self::unique_locs.

analyzed_locs: u32

The number of lines we ran through the PDG construction. This is higher than unique LoCs, because we need to analyze some functions multiple times, due to monomorphization and calls tring differences.

analyzed_functions: u32

Number of functions that correspond to [Self::analyzed_locs]

inlinings_performed: u32

How many times we inlined functions. This will be higher than Self::analyzed_functions because sometimes the callee PDG is served from the cache.

construction_time: Duration

How long it took to create this PDG

conversion_time: Duration

How long it took to calculate markers and otherwise set up the pdg